webapp/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-dotenv==1.0.1
google-genai==1.14.0
requests==2.32.3
aiofiles==24.1.0
sse-starlette==3.3.2
